Change Integer 'limb' size to 32-bits

+++ lib/types/bytes.h
@@ -8,35 +8,34 @@
 namespace MathSolver
 {
 
-// Adds a pair of single bytes, stores the resultant uint8_t at the location specified by
+// Adds a pair of quad bytes, stores the resultant uint32_t at the location specified by
 // res, and returns the overflow. 
-uint8_t addByte2(uint8_t* res, uint8_t a, uint8_t b);
+uint32_t addByte2(uint32_t* res, uint32_t a, uint32_t b);
 
-// Adds a triple of single bytes, stores the resultant uint8_t at the location specified by
+// Adds a triple of quad bytes, stores the resultant uint32_t at the location specified by
 // res, and returns the overflow. 
-uint8_t addByte3(uint8_t* res, uint8_t a, uint8_t b, uint8_t c);
+uint32_t addByte3(uint32_t* res, uint32_t a, uint32_t b, uint32_t c);
 
-// Compares two byte arrays by highest non-zero bit. Returns 1 if a is greater than b,
+// Compares two quad byte arrays by highest non-zero bit. Returns 1 if a is greater than b,
 // 0 if a equals b, and -1 if a is less than b.
-int cmpBytes(uint8_t* a, size_t alen, uint8_t* b, size_t blen);
+int cmpBytes(uint32_t* a, size_t alen, uint32_t* b, size_t blen);
 
-// Returns the value of a given bit in a byte array. Assertion: bit < 8 * len.
-uint8_t getBit(uint8_t* x, size_t len, size_t bit);
+// Returns the value of a given bit in a quad byte array. Assertion: bit < 8 * len.
+uint32_t getBit(uint32_t* x, size_t len, size_t bit);
 
 // Returns the highest non-zero bit. Returns 0 if all bytes are zero.
-size_t highestNonZeroBit(uint8_t* x, size_t len);
+size_t highestNonZeroBit(uint32_t* x, size_t len);
 
-// Returns one higher than the index of the last non-zero byte in a given byte array.
+// Returns one higher than the index of the last non-zero quad byte in a given quad byte array.
 // (Returns 0 if all bytes are zero.)
-size_t highestNonZeroByte(uint8_t* x, size_t len);
-
+size_t highestNonZeroByte(uint32_t* x, size_t len);
 
 // Returns true if all bytes within the specified range are equal to 0. Returns false
 // otherwise. Assertion: high >= low.
-bool rangeIsEmpty(uint8_t* low, uint8_t* high);
+bool rangeIsEmpty(uint32_t* low, uint32_t* high);
 
-// Sets a given bit in a byte array. Assertion: bit < 8 * len.
-void setBit(uint8_t* x, size_t len, size_t bit, bool value);
+// Sets a given bit in a quad byte array. Assertion: bit < 8 * len.
+void setBit(uint32_t* x, size_t len, size_t bit, bool value);
 
 } // END MathSolver namespace
